Fix user undefined error
#12
