--- components/UserInfo.js
@@ -1,25 +1,0 @@
-import React, { useState, useEffect } from 'react';
-
-async function getUserInfo() {
-  try {
-    const response = await fetch('/.auth/me');
-    const payload = await response.json();
-    const { clientPrincipal } = payload;
-    return clientPrincipal;
-  } catch (error) {
-    console.error('No profile could be found');
-    return undefined;
-  }
-}
-
-export default function userInfo() {
-  const [userInfo, setUserInfo] = useState();
-
-  useEffect(() => {
-    (async () => {
-      setUserInfo(await getUserInfo());
-    })();
-  }, []);
-
-  return `${userInfo && userInfo.userDetails}`;
-}
