+++ components/navbar.js
@@ -5,54 +5,57 @@ import {
   InteractionOutlined,
   UserOutlined,
 } from '@ant-design/icons';
-import React from 'react';
-import UserInfo from './UserInfo';
+import React, { useState, useEffect } from 'react';
 
 const { SubMenu } = Menu;
 
-class App extends React.Component {
-  state = {
-    current: undefined,
-  };
+export default function App() {
+  const [userInfo, setUserInfo] = useState();
 
-  handleClick = (e) => {
-    console.log('click ', e);
-    this.setState({ current: e.key });
-  };
+  useEffect(() => {
+    (async () => {
+      setUserInfo(await getUserInfo());
+    })();
+  }, []);
 
-  render() {
-    const { current } = this.state;
-    return (
-      <Menu
-        onClick={this.handleClick}
-        selectedKeys={[current]}
-        mode="horizontal"
-      >
-        <Menu.Item key="portfolio" icon={<HomeOutlined />}>
-          <a href="/authenticated/portfolio/">Portfolio</a>
-        </Menu.Item>
-        <Menu.Item key="performance" icon={<AreaChartOutlined />}>
-          <a href="/authenticated/performance/">Performance</a>
-        </Menu.Item>
-        <Menu.Item key="actions" icon={<InteractionOutlined />}>
-          <a href="/authenticated/actions/">Actions</a>
-        </Menu.Item>
+  async function getUserInfo() {
+    try {
+      const response = await fetch('/.auth/me');
+      const payload = await response.json();
+      const { clientPrincipal } = payload;
+      return clientPrincipal;
+    } catch (error) {
+      console.error('No profile could be found');
+      return undefined;
+    }
+  }
+
+  return (
+    <Menu mode="horizontal">
+      <Menu.Item key="portfolio" icon={<HomeOutlined />}>
+        <a href="/authenticated/portfolio/">Portfolio</a>
+      </Menu.Item>
+      <Menu.Item key="performance" icon={<AreaChartOutlined />}>
+        <a href="/authenticated/performance/">Performance</a>
+      </Menu.Item>
+      <Menu.Item key="actions" icon={<InteractionOutlined />}>
+        <a href="/authenticated/actions/">Actions</a>
+      </Menu.Item>
+      {userInfo && (
         <SubMenu
           icon={<UserOutlined />}
           className="ml-auto"
           key="SubMenu"
-          title={<UserInfo />}
+          title={userInfo && userInfo.userDetails}
         >
           <Menu.Item key="logout">
-            <a href="/.auth/logout?post_logout_redirect_uri=/">logout</a>
+            <a href="/.auth/logout?post_logout_redirect_uri=/">Logout</a>
           </Menu.Item>
           <Menu.Item key="settings">
             <a href="/authenticated/settings">Settings</a>
           </Menu.Item>
         </SubMenu>
-      </Menu>
-    );
-  }
+      )}
+    </Menu>
+  );
 }
-
-export default App;
